sales-api/requirements.txt
fastapi~=0.88.0
openai~=0.27.8
uvicorn~=0.22.0
pinecone-client~=2.2.2
openai~=0.27.8
python-dotenv~=1.0.0